cdk/requirements.txt
aws-cdk.cdk==1.0.0
aws-cdk.core==1.0.0
aws-cdk.aws_ec2==1.0.0
aws-cdk.aws_ecs==1.0.0
aws-cdk.aws_ecs_patterns==1.0.0
aws-cdk.aws_elasticloadbalancingv2==1.0.0
aws-cdk.aws_logs==1.0.0
aws-cdk.aws_servicediscovery==1.0.0
